% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyalert.R
\name{close_alert}
\alias{close_alert}
\title{The callback functions are not called when the modal is forced to close.}
\usage{
close_alert()
}
\description{
The callback functions are not called when the modal is forced to close.
}
