% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyalert.R
\name{dismissalert}
\alias{dismissalert}
\title{Dismiss one or more popup messages}
\usage{
dismissalert(n = NULL)
}
\arguments{
\item{n}{Number of popup messages to dismiss. If set to \code{NULL}, then
all modals will be dismissed; default is \code{NULL}}
}
\description{
Dismiss one or more popup messages
}
