% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run shinyalert example}
\usage{
runExample()
}
\description{
Launch an example Shiny app that shows how easy it is to
create modals with \code{shinyalert}.\cr\cr
The demo app is also
\href{https://daattali.com/shiny/shinyalert-demo/}{available online}
to experiment with.
}
