% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyalert-package.R
\docType{package}
\name{shinyalert-package}
\alias{shinyalert-package}
\alias{_PACKAGE}
\title{shinyalert: Easily Create Pretty Popup Messages (Modals) in 'Shiny'}
\description{
Easily create pretty popup messages (modals) in 'Shiny'. A modal can contain text, images, OK/Cancel buttons, an input to get a response from the user, and many more customizable options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/shinyalert}
  \item \url{https://daattali.com/shiny/shinyalert-demo/}
  \item Report bugs at \url{https://github.com/daattali/shinyalert/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID}) (R interface)

Authors:
\itemize{
  \item Tristan Edwards (sweetalert library)
}

Other contributors:
\itemize{
  \item Zhengjia Wang [contributor]
}

}
\keyword{internal}
