% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyalert.R
\name{useShinyalert}
\alias{useShinyalert}
\title{Set up a Shiny app to use shinyalert}
\usage{
useShinyalert(rmd = FALSE, force = FALSE)
}
\arguments{
\item{rmd}{Set this to \code{TRUE} if using \code{shinyalert}
inside an interactive R markdown document. The YAML of the Rmd file must have
\code{runtime: shiny}.}

\item{force}{Set to \code{TRUE} to load the {shinyalert} scripts. If \code{FALSE} (default),
you will get a warning saying this function is not required.}
}
\value{
Scripts that \code{shinyalert} requires that are automatically
inserted to the app's \code{<head>} tag.
}
\description{
\strong{This function is no longer needed.}\cr\cr
When a {shinyalert} message is shown for the first time,
the required scripts are automatically inserted to the Shiny app. If you want to pre-load
these scripts into the app for performance reasons, you can call this function anywhere
in the UI and use \code{force = TRUE}.\cr\cr
Note that it is no longer needed to use this function.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyalert)

  shinyApp(
    ui = fluidPage(
      useShinyalert(force = TRUE),  # Set up shinyalert
      actionButton("btn", "Click me")
    ),
    server = function(input, output) {
      observeEvent(input$btn, {
        # Show a simple modal
        shinyalert(title = "You did it!", type = "success")
      })
    }
  )
}
}
\seealso{
\code{\link[shinyalert]{shinyalert}}
}
