% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeAlert.R
\name{closeAlert}
\alias{closeAlert}
\title{Close any open shinyalert modals}
\usage{
closeAlert()
}
\description{
Close any shinyalert modals that are currently open. The callback functions
of the modal are not called when the modal is closed this way.
}
