% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyalert.R
\name{closeAlert}
\alias{closeAlert}
\title{TODO
The callback functions are not called when the modal is forced to close.}
\usage{
closeAlert()
}
\description{
TODO
The callback functions are not called when the modal is forced to close.
}
