% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyalert.R
\name{closeAlert}
\alias{closeAlert}
\title{Close a shinyalert popup message}
\usage{
closeAlert(num = 0)
}
\arguments{
\item{num}{Number of popup messages to close. If set to 0 (default) then all
messages are closed. This is only useful if you have multiple popups queued up.}
}
\description{
Close a shinyalert popup message
}
